requirements.txt
pandas>=2.2.0
pyarrow>=15.0.0
requests>=2.31.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dateutil>=2.9.0.post0
